python/pyproject.toml
[tool.poetry]
name = "segway-connect-system-destination-otlp"
version = "0.1.0"
description = ""
authors = ["Ryan Faircloth <ryan@dss-i.com>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.9"
Jinja2 = "^3.1.2"
PyYAML = "^6.0.1"
furl = "^2.1.3"
python-dotenv = "^1.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
